refactor(init): promote good practices; use site alias
