--- packages/create-docusaurus/templates/classic-typescript/src/components/HomepageFeatures.module.css
@@ -1,1 +1,0 @@
-../../../classic/src/components/HomepageFeatures.module.css
