+++ packages/create-docusaurus/templates/classic-typescript/src/components/HomepageFeatures/styles.module.css
@@ -1,0 +1,1 @@
+../../../../classic/src/components/HomepageFeatures/styles.module.css
